pyproject.toml
[tool.poetry]
name = "fipe"
version = "0.1.0"
description = ""
authors = ["doug-pires <dpiresmartins@yahoo.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
async-generator = "1.10"
polars = "^0.17.2"
beautifulsoup4 = "^4.12.2"
selenium = "^4.8.3"
webdriver-manager = "^3.8.5"
pyyaml = "^6.0"
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.17"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pyspark = "^3.4.0"
delta-spark = "^2.4.0"
isort = "^5.12.0"
types-pyyaml = "^6.0.12.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = ["--strict-markers", "--strict-config", "-ra"]
testpaths = ["tests"]
markers = [
    { name = "extract", description = "tests for extraction code" }
]

[tool.black]
line-length = 88